refactor(hook): split hook_cmd.rs into modular directory structure
Split the monolithic 1127-line hook_cmd.rs into 8 focused modules under
src/hook/ for better maintainability and code review:

- mod.rs: entry point, chain parsing, orchestrator
- git.rs: git + GitHub CLI rewriters
- cargo.rs: cargo rewriter
- files.rs: file ops (cat, grep, ls, tree, find, diff, head) + network
- js_ts.rs: JS/TS tooling (vitest, tsc, eslint, prettier, playwright, prisma, pnpm)
- containers.rs: Docker + kubectl rewriters
- python.rs: pytest, ruff, pip/uv rewriters
- go.rs: go test/build/vet + golangci-lint rewriters
- helpers.rs: shared utility functions (replace_prefix, starts_with_any)

All 71 hook tests passing. No behavioral changes.
